+++ smartdeals-service/src/main/java/data/smartdeals_service/services/forumServices/QuestionService.java
@@ -2,10 +2,15 @@ package data.smartdeals_service.services.forumServices;
 
 import data.smartdeals_service.dto.forum.QuestionDTO;
 import data.smartdeals_service.dto.forum.QuestionResponseDTO;
+import data.smartdeals_service.dto.forum.QuestionStatusDTO;
 import data.smartdeals_service.dto.forum.UpdateQuestionDTO;
+import data.smartdeals_service.dto.promotion.PromotionStatusDTO;
 import data.smartdeals_service.helpers.FileUploadAvata;
 import data.smartdeals_service.helpers.GlobalConstant;
+import data.smartdeals_service.models.comment.Comment;
 import data.smartdeals_service.models.forum.*;
+import data.smartdeals_service.models.promotion.Promotion;
+import data.smartdeals_service.repository.commentRepositories.CommentRepository;
 import data.smartdeals_service.repository.forumRepositories.QuestionImageRepository;
 import data.smartdeals_service.repository.forumRepositories.QuestionRepository;
 import data.smartdeals_service.repository.forumRepositories.QuestionViewRepository;
@@ -32,6 +37,7 @@ public class QuestionService {
     private final QuestionImageRepository questionImageRepository;
     private final QuestionVoteRepository questionVoteRepository;
     private final QuestionViewRepository questionViewRepository;
+    private final CommentRepository commentRepository;
     private final FileUploadAvata fileUploadAvata;
     private String subFolder = "QuestionsImage";
     private String urlImage = GlobalConstant.rootUrl
@@ -43,9 +49,9 @@ public class QuestionService {
         question.setAuthor(questionDTO.getAuthor());
         question.setTitle(questionDTO.getTitle());
         question.setContent(questionDTO.getContent());
-        question.setTopic(questionDTO.getTopic());
         question.setTag(questionDTO.getTag());
         question.setRolePost(questionDTO.getRolePost());
+        question.setStatus(questionDTO.getStatus());
 
         if (questionDTO.getCategory() != null) {
             List<CategoryForum> categories = questionDTO.getCategory().stream()
@@ -95,7 +101,6 @@ public List<QuestionResponseDTO> findAllQuestions() {
         }
         questionById.setTitle(updateQuestionDTO.getTitle());
         questionById.setContent(updateQuestionDTO.getContent());
-        questionById.setTopic(updateQuestionDTO.getTopic());
         questionById.setTag(updateQuestionDTO.getTag());
         questionById.setUpdatedAt(LocalDateTime.now());
 
@@ -145,14 +150,30 @@ public List<QuestionResponseDTO> findAllQuestions() {
 
     public void deleteQuestion(Long id) {
         Optional<Question> questionById = questionRepository.findById(id);
-        List<QuestionImage> imagesDelete = questionById.get().getQuestionImage();
-        if (imagesDelete.size()>0) {
-            for (QuestionImage image : imagesDelete) {
-                fileUploadAvata.deleteImage(image.getImageUrl().substring(GlobalConstant.rootUrl.length()));
-                questionImageRepository.delete(image);
+
+        // Kiểm tra nếu câu hỏi tồn tại
+        if (questionById.isPresent()) {
+            List<Comment> commentsToDelete = commentRepository.findCommentsByQuestionId(id);
+
+            // Nếu có bình luận, xóa chúng
+            if (!commentsToDelete.isEmpty()) {
+                for (Comment data : commentsToDelete) {
+                    commentRepository.delete(data);  // Xóa bình luận
+                }
             }
+
+            List<QuestionImage> imagesDelete = questionById.get().getQuestionImage();
+            if (imagesDelete.size() > 0) {
+                for (QuestionImage image : imagesDelete) {
+                    fileUploadAvata.deleteImage(image.getImageUrl().substring(GlobalConstant.rootUrl.length()));
+                    questionImageRepository.delete(image);
+                }
+            }
+            questionRepository.deleteById(id);
+        }else {
+            // Xử lý nếu câu hỏi không tồn tại
+            throw new RuntimeException("Câu hỏi không tồn tại!");
         }
-        questionRepository.deleteById(id);
     }
 
 
@@ -226,4 +247,10 @@ public List<QuestionResponseDTO> findAllQuestions() {
         // Nếu đã xem, không làm gì thêm
     }
 
+    public Question closeQuestionStatus(Long id, QuestionStatusDTO status) {
+        Question question = questionRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("question not found with id: " + id));
+        question.setStatus(status.getStatus());
+        return questionRepository.save(question);
+    }
 }
